debug

+++ src/MLEC/models/SpanEmoB2B.py
@@ -86,7 +86,7 @@ class SpanEmoB2B(MLECModel):
             target_attention_masks.to(device),
         )
 
-        all_label_input_ids = all_label_input_ids.long().to(device)
+        all_label_input_ids = all_label_input_ids.to(device)
         label_idxs = label_idxs[0].long().to(device)
 
         outputs = self.model(
